+++ Engine/vortex.hh
@@ -9,6 +9,7 @@ namespace Common {
 
 /**
  * @brief Actor that destroys everything from adjacent water hexes.
+ * Removes itself from the board immediately after action.
  */
 class Vortex : public Actor
 {
@@ -18,6 +19,10 @@ public:
     * @param typeId the id of the vortex.
     */
     Vortex(int typeId);
+    /**
+     * @copydoc Common::Actor::move()
+     */
+    virtual void move( std::shared_ptr<Common::Hex> to );
     /**
      * @copydoc Common::Actor::doAction()
      */